services/perception-api/requirements.txt
# =============================================================================
# SAM3 Perception API - Python Dependencies
# =============================================================================

# Core ML
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0

# Image processing
pillow>=10.0.0
opencv-python-headless>=4.8.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async HTTP
httpx>=0.26.0
aiofiles>=23.2.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Caching
redis>=5.0.0

# Logging & Observability
structlog>=24.1.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
prometheus-client>=0.19.0

# CLI
typer>=0.9.0
rich>=13.7.0
